refactor: PR #344 Round 11 리뷰 코멘트 반영 (cn() 적용, 탭 유틸 분리, useMemo, 타입 중앙화)
