+++ src/components/ui/tabs/TabsPill.tsx
@@ -3,11 +3,7 @@
 import { useMemo, type ReactElement } from 'react';
 import { cn } from '@/lib/cn';
 import { useTabs } from './useTabs';
-
-interface TabItem<T extends string> {
-    value: T;
-    label: string;
-}
+import type { TabItem } from './utils/tabIds';
 
 interface TabsPillProps<T extends string> {
     tabs: readonly TabItem<T>[];
@@ -28,11 +24,7 @@ export function TabsPill<T extends string>({
     withHomeEnd = false,
     className,
 }: TabsPillProps<T>): ReactElement {
-    // useMemo 반환 타입이 T[]로 추론되므로 readonly 제약을 위해 캐스트한다.
-    const values = useMemo(
-        () => tabs.map(t => t.value),
-        [tabs]
-    ) as readonly T[];
+    const values = useMemo((): readonly T[] => tabs.map(t => t.value), [tabs]);
     const { getTabProps } = useTabs({
         tabs: values,
         activeTab,
