+++ src/components/ui/tabs/index.ts
@@ -1,3 +1,4 @@
-export { useTabs, buildTabId, buildPanelId } from './useTabs';
+export { useTabs } from './useTabs';
+export { buildTabId, buildPanelId, type TabItem } from './utils/tabIds';
 export { TabsUnderline } from './TabsUnderline';
 export { TabsPill } from './TabsPill';
